gcloud/functions/event_handler/requirements.txt
functions-framework==3.*
google-cloud-bigquery>=3.18.0,<=4
